+++ src/client/render/types/UnitType.ts
@@ -24,6 +24,7 @@ export const UT_TRAIN = "Train" as const;
 export const UT_BOMBER = "Bomber" as const;
 export const UT_SUBMARINE = "Submarine" as const;
 export const UT_CARRIER = "Carrier" as const;
+export const UT_PARATROOPER = "Paratrooper" as const;
 
 // Structures
 export const UT_CITY = "City" as const;
@@ -105,4 +106,5 @@ export const ALL_UNIT_TYPES = [
   UT_BOMBER,
   UT_SUBMARINE,
   UT_CARRIER,
+  UT_PARATROOPER,
 ] as const;